implemented user permissions update

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -2,12 +2,22 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          user: {
+            connect: {
+              id: ctx.request.userId,
+            },
+          },
           ...args,
         },
       },
@@ -32,9 +42,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1 find theitem
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     // 2 check if they own that item or check permission
     // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(
+      permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+    )
+    if(!ownsItem && !hasPermissions){
+      throw new Error('You are not allowed to do that!')
+    }
     // 3 delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -110,7 +127,6 @@ const Mutations = {
       throw new Error('Yours passwords dont match');
     }
     // 2 check if it is legit resettoken
-    console.log('EXPIRY CALC', Date.now() - 360000);
     const [user] = await ctx.db.query.users({
       where: {
         resetToken: args.resetToken,
@@ -141,6 +157,38 @@ const Mutations = {
     });
     // 8 return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1 check if logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in');
+    }
+    // 2 query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info
+    );
+    console.log('THIS IS THE USER!', currentUser);
+    // 3 check if they have permission to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4 update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info
+    );
   },
 };
